pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "criteria-bge-hpo"
version = "0.1.0"
description = "DSM-5 Criteria Matching using Natural Language Inference with BGE reranker"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "CVR Lab 308" }]
keywords = ["nlp", "nli", "bert", "dsm5", "mental-health", "mlflow", "optuna"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "torch>=2.0.0",
  "transformers>=4.48.0",
  "mlflow>=2.5.0",
  "optuna>=3.4.0",
  "psycopg2-binary>=2.9.9",
  "hydra-core>=1.3.0",
  "omegaconf>=2.3.0",
  "rich>=13.0.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "tqdm>=4.65.0",
  "nlpaug>=1.1.11",
  "textattack>=0.3.10",
  "nltk>=3.8.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.6",
  "black>=24.3",
  "mypy>=1.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
